backend/requirement.txt
anthropic
termcolor
Pillow
sqlalchemy